# specs/001-ai-payment-agent/contracts/openapi.yaml
openapi: "3.0.3"
info:
  title: AI Freelancer Payment Agent API
  version: "1.0.0"
  description: |
    Backend API for the AI Freelancer Payment Agent.
    All endpoints except /auth/register and /auth/login require Bearer JWT authentication.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://your-backend.onrender.com
    description: Production (Render)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LineItem:
      type: object
      required: [description, quantity, unit_price]
      properties:
        description: { type: string }
        quantity: { type: number, minimum: 0.01 }
        unit_price: { type: number, minimum: 0 }
        total: { type: number, readOnly: true }
    InvoiceStatus:
      type: string
      enum: [DRAFT, PENDING_APPROVAL, APPROVED, REJECTED, SETTLED, OVERDUE]
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        line_items: { type: array, items: { $ref: '#/components/schemas/LineItem' } }
        total_amount: { type: number }
        currency: { type: string, default: USDC }
        due_date: { type: string, format: date }
        payment_terms: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/InvoiceStatus' }
        freelancer_id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateInvoiceRequest:
      type: object
      required: [title, line_items, due_date, client_id]
      properties:
        title: { type: string }
        line_items: { type: array, items: { $ref: '#/components/schemas/LineItem' }, minItems: 1 }
        due_date: { type: string, format: date }
        payment_terms: { type: string }
        client_id: { type: string, format: uuid }
    AuditEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        invoice_id: { type: string, format: uuid }
        actor_id: { type: string, format: uuid }
        prev_status: { type: string, nullable: true }
        new_status: { type: string }
        note: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
...
security:
  - BearerAuth: []
paths:
...
  /api/invoices:
    get:
      summary: List invoices for current user
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/InvoiceStatus' }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Invoice' } }
                  total: { type: integer }
                  page: { type: integer }

    post:
      summary: Create new invoice (FREELANCER only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInvoiceRequest' }
      responses:
        '201': { description: Invoice created, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
        '403': { description: Only freelancers can create invoices }
        '422': { description: Validation error }
  /api/invoices/{id}:
    get:
      summary: Get invoice details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Invoice details, content: { application/json: { schema: { $ref: '#/components/schemas/Invoice' } } } }
        '404': { description: Invoice not found }

    patch:
      summary: Update DRAFT invoice (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInvoiceRequest' }
      responses:
        '200': { description: Updated invoice }
        '403': { description: Cannot edit non-DRAFT invoice or not owner }

    delete:
      summary: Delete DRAFT invoice (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Invoice deleted }
        '403': { description: Cannot delete non-DRAFT invoice }
  /api/invoices/{id}/submit:
    post:
      summary: Submit DRAFT invoice for approval
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Invoice submitted, status → PENDING_APPROVAL }
        '403': { description: Not the invoice owner }
        '422': { description: Invoice is not in DRAFT status }
  /api/invoices/{id}/approve:
    post:
      summary: Approve invoice (assigned CLIENT only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Invoice approved, status → APPROVED }
        '403': { description: Not the assigned client }
  /api/invoices/{id}/reject:
    post:
      summary: Reject invoice with reason (assigned CLIENT only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 10 }
      responses:
        '200': { description: Invoice rejected, status → REJECTED }
  /api/invoices/{id}/audit:
    get:
      summary: Get invoice audit trail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Audit trail
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEntry' }
...